Update on "Refactor CssSelector into class"
[ghstack-poisoned]
