+++ packages/core/src/locators/byAttribute.ts
@@ -1,6 +1,7 @@
 import { escapeName, escapeValue } from '../utils/escapeUtil';
+import { CssLocator } from './CssLocator';
 import { LocatorRelativePosition } from './LocatorRelativePosition';
-import { CssLocator, PartLocatorType } from './PartLocatorType';
+import { PartLocatorType } from './PartLocatorType';
 
 export type ByAttributeSource = {
   _id: 'byAttribute';
