+++ packages/core/src/locators/byCssClass.ts
@@ -1,6 +1,7 @@
 import { escapeCssClassName } from '../utils/escapeUtil';
+import { CssLocator } from './CssLocator';
 import { LocatorRelativePosition } from './LocatorRelativePosition';
-import { CssLocator, PartLocatorType } from './PartLocatorType';
+import { PartLocatorType } from './PartLocatorType';
 
 export type ByCssClassSource = {
   _id: 'byCssClass';
